% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_pick_2.R
\name{anovapick.2}
\alias{anovapick.2}
\title{Pick Non-Zero Correlations in a Lossy Matrix}
\usage{
anovapick.2(matrix, y_vector, alpha = 0.025)
}
\arguments{
\item{matrix, }{matrix of lossy data}

\item{y_vector, }{vector of output y's}

\item{alpha, }{default is 0.025, lower tail probability of not rejecting the null when it is indeed true (r=0)}
}
\description{
Pick Non-Zero Correlations in a Lossy Matrix
}
