% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_loss.R
\name{corr.loss}
\alias{corr.loss}
\title{Generate vector of correlations between x matrix and vector of y's}
\usage{
corr.loss(matrix, y_vector = NA)
}
\arguments{
\item{matrix, }{matrix of lossy data}

\item{y_vector, }{vector of output y's, if no y_vector input defaults to using the last column of matrix}
}
\description{
Generate vector of correlations between x matrix and vector of y's
}
