% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_pick.R
\name{fish.pick}
\alias{fish.pick}
\title{Pick Non-Zero Correlations in a Lossy Matrix}
\usage{
fish.pick(matrix, y_vector, alpha = 0.025)
}
\arguments{
\item{matrix, }{matrix of lossy data}

\item{y_vector, }{vector of output y's}

\item{alpha, }{default is 0.025, lower tail probability of not rejecting the null when it is indeed true (r=0)}
}
\description{
Pick Non-Zero Correlations in a Lossy Matrix
}
