% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_rank.R
\name{fish.rank}
\alias{fish.rank}
\title{This function generates ranks, either jaccard coefficients (default) or standardized mean rank (type='standmean') for a vector of correlations (vector) over a set of confidence intervals (conf).}
\usage{
fish.rank(vector, matrix, full_matrix, y_vector, conf = (1:99)/100,
  m = NA, true_match = NA, type = "jaccard", z = 2,
  lastcol_matrix_y = F, lastcol_full_y = F, keep_matches = F)
}
\arguments{
\item{vector, }{no default, vector of correlations}

\item{matrix, }{lossy matrix from which the the correlations were generated}

\item{full_matrix, }{full matrix from which the lossy matrix was generated}

\item{y_vector, }{vector of y's, defaults to the last column of a matrix if either lastcol_matrix_y or lastcol_full_y is set to TRUE}

\item{conf, }{default=(1:99)/100. Input index of correlations as probabilities.}

\item{m, }{default=10 percent of total variables. Number of matches to choose, integer}

\item{type, }{default='jaccard'. Only other option is 'standmean'.}

\item{z, }{default=2, factor by which to make the set compared in jaccard coefficient larger than m. Should be an integer.}

\item{lastcol_matrix_y, }{default=F, TRUE/FALSE that determines if the last column of matrix is the y_vector.}

\item{lastcol_full_y, }{default=F, TRUE/FALSE that determines if the last column of the full_matrix is the y_vector.}

\item{keep_matches, }{default=F, if true will bind the matches from every round into a list with the ranks}
}
\description{
This function generates ranks, either jaccard coefficients (default) or standardized mean rank (type='standmean') for a vector of correlations (vector) over a set of confidence intervals (conf).
}
