% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uafs_pick.R
\name{uafs.pick}
\alias{uafs.pick}
\title{Pick significant relationships between output and columns in a data frame}
\usage{
uafs.pick(data, output, alpha = 0.025)
}
\arguments{
\item{data, }{matrix with output data}

\item{output, }{output variable, of type either character or numeric}

\item{alpha, }{default = 0.025,  lower tail probability of not rejecting the null when it is indeed true (r=0)}
}
\description{
Pick significant relationships between output and columns in a data frame
}
